last changes

+++ src/eye/admin/table/EyeTableModel.java
@@ -4,48 +4,77 @@
  */
 package eye.admin.table;
 
-import com.db4o.ObjectContainer;
-import eye.server.manager.impl.DBManagerBasicImpl;
-import eye.core.model.RemoteSource;
 import java.util.List;
-import java.util.Vector;
 import javax.swing.table.DefaultTableModel;
+import com.db4o.ObjectContainer;
+import com.db4o.ext.Db4oRecoverableException;
+import eye.admin.MainFrame;
+import eye.core.model.Image;
+import eye.core.model.Place;
+import eye.core.model.RemoteSource;
+import eye.server.manager.DBManagerBasicImpl;
+import java.util.Hashtable;
+import java.util.Map.Entry;
+import java.util.Vector;
 
 /**
  *
  * @author stream
- * @version $Id: EyeTableModel.java 71 2010-07-08 03:50:40Z spr1ng $
+ * @version $Id: EyeTableModel.java 109 2010-07-19 00:22:16Z stream $
  */
 public class EyeTableModel extends DefaultTableModel {
 
-    private static eye.server.manager.impl.DBManagerBasicImpl dbm = new DBManagerBasicImpl();
+    private static DBManagerBasicImpl dbm = new DBManagerBasicImpl();
     private ObjectContainer db = dbm.getContainer();
+    private Hashtable<String, String> cache = new Hashtable<String, String>();
 
     public EyeTableModel() {
         super(new Object[]{"Idx", "URL", " "}, 0);
     }
 
-    public void clear() {
-        getDataVector().removeAllElements();
-        fireTableDataChanged();
+    public void cleanDatabase() {
+        Vector<Vector> dv = getDataVector();
+        if (MainFrame.getRemoteSourceClass().equals("Images")) {
+            for (Vector<String> row : dv) {
+                RemoteSource img = new Image(row.get(1));
+                delete(img);
+            }
+        } else {
+            for (Vector<String> row : dv) {
+                RemoteSource place = new Place(row.get(1));
+                delete(place);
+            }
+        }
+        clearTable();
     }
 
     public void addEmptyRow() {
-        Vector row = new Vector();
-        row.add(" ");
-        row.add("http://");
-        row.add(" ");
-        addRow(row);
+        Object rowData[] = new Object[]{"  ", "http://", " "};
+        addRow(rowData);
+        fireTableDataChanged();
+    }
+
+    @Override
+    public void setValueAt(Object aValue, int row, int column) {
+        super.setValueAt(aValue, row, column);
+        cache(aValue.toString());
     }
 
     public void delete(RemoteSource rs) {
-        db.delete(select(rs));
-        db.commit();
+        RemoteSource srs = select(rs);
+        if (db != null && srs != null) {
+            db.delete(select(rs));
+            db.commit();
+        }
     }
 
     private RemoteSource select(RemoteSource rs) {
-        List<RemoteSource> remoteSourceList = db.queryByExample(rs);
-        return remoteSourceList.get(0);
+        try {
+            List<RemoteSource> remoteSourceList = db.queryByExample(rs);
+            return remoteSourceList.get(0);
+        } catch (Db4oRecoverableException recex) {
+            return null;
+        }
     }
 
     @Override
@@ -53,11 +82,35 @@ public class EyeTableModel extends DefaultTableModel {
         return true;
     }
 
-    public void cache() {
-        System.out.println("STUB");
+    public void cache(String remoteSourceURL) {
+        cache.put(remoteSourceURL, MainFrame.getRemoteSourceClass());
+    }
+
+    public void commitNewData() {
+        RemoteSource rs = null;
+        if (db != null) {
+            if (MainFrame.getRemoteSourceClass().equals("Images")) {
+                for (Entry e : cache.entrySet()) {
+                    rs = new Image(e.getKey().toString());
+                    db.store(rs);
+                }
+            } else {
+                for (Entry e : cache.entrySet()) {
+                    rs = new Place(e.getKey().toString());
+                    db.store(rs);
+                }
+            }
+            commit();
+        }
+        cache.clear();
     }
 
     public void commit() {
-        dbm.getContainer().commit();
+        db.commit();
+    }
+
+    public void clearTable() {
+        getDataVector().removeAllElements();
+        fireTableDataChanged();
     }
 }
